moved distOnRead method to CigarUtils with no other change
